backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.9.1
httpx==0.27.2
requests==2.32.3
python-dotenv==1.0.1
stripe==11.1.0
twilio==9.3.7
postmarker==1.0
python-multipart==0.0.12
beautifulsoup4==4.12.3
lxml==5.3.0
PyJWT==2.9.0
cryptography==43.0.1
APScheduler==3.10.4
sentry-sdk[fastapi]==2.17.0
